Pipelines/Azure_Functions/python_durable_function/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
azure-ai-formrecognizer==1.0.0b3
azure-core==1.6.0
azure-functions==1.2.1
azure-functions-durable==1.0.0b6
certifi==2020.6.20
chardet==3.0.4
filetype==1.0.7
furl==2.1.0
idna==2.9
install==1.3.1
isodate==0.6.0
msrest==0.6.16
multidict==4.7.6
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.2.0.34
orderedmultidict==1.0.1
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.9
yarl==1.4.2
